correction_server/requirements.txt
Flask==3.0.3
gramformer==1.0
pdfminer.six==20240706
